+++ src/opengeodeweb_viewer/rpc/mesh/polyhedra/polyhedra_protocols.py
@@ -9,9 +9,9 @@ from opengeodeweb_viewer.utils_functions import get_schemas_dict, validate_schem
 from opengeodeweb_viewer.rpc.mesh.mesh_protocols import VtkMeshView
 
 
-class VtkMeshPolyhedronsView(VtkMeshView):
-    mesh_polyhedrons_prefix = "opengeodeweb_viewer.mesh.polyhedrons."
-    mesh_polyhedrons_schemas_dict = get_schemas_dict(
+class VtkMeshPolyhedraView(VtkMeshView):
+    mesh_polyhedra_prefix = "opengeodeweb_viewer.mesh.polyhedra."
+    mesh_polyhedra_schemas_dict = get_schemas_dict(
         os.path.join(os.path.dirname(__file__), "schemas")
     )
 
@@ -19,23 +19,23 @@ class VtkMeshPolyhedronsView(VtkMeshView):
         super().__init__()
 
     @exportRpc(
-        mesh_polyhedrons_prefix + mesh_polyhedrons_schemas_dict["visibility"]["rpc"]
+        mesh_polyhedra_prefix + mesh_polyhedra_schemas_dict["visibility"]["rpc"]
     )
-    def setMeshPolyhedronsVisibility(self, params):
+    def setMeshPolyhedraVisibility(self, params):
         validate_schema(
             params,
-            self.mesh_polyhedrons_schemas_dict["visibility"],
-            self.mesh_polyhedrons_prefix,
+            self.mesh_polyhedra_schemas_dict["visibility"],
+            self.mesh_polyhedra_prefix,
         )
         id, visibility = params["id"], params["visibility"]
         self.SetVisibility(id, visibility)
 
-    @exportRpc(mesh_polyhedrons_prefix + mesh_polyhedrons_schemas_dict["color"]["rpc"])
-    def setMeshPolyhedronsColor(self, params):
+    @exportRpc(mesh_polyhedra_prefix + mesh_polyhedra_schemas_dict["color"]["rpc"])
+    def setMeshPolyhedraColor(self, params):
         validate_schema(
             params,
-            self.mesh_polyhedrons_schemas_dict["color"],
-            self.mesh_polyhedrons_prefix,
+            self.mesh_polyhedra_schemas_dict["color"],
+            self.mesh_polyhedra_prefix,
         )
         id, red, green, blue = (
             params["id"],
@@ -46,27 +46,27 @@ class VtkMeshPolyhedronsView(VtkMeshView):
         self.SetColor(id, red, green, blue)
 
     @exportRpc(
-        mesh_polyhedrons_prefix
-        + mesh_polyhedrons_schemas_dict["vertex_attribute"]["rpc"]
+        mesh_polyhedra_prefix
+        + mesh_polyhedra_schemas_dict["vertex_attribute"]["rpc"]
     )
-    def setMeshPolyhedronsVertexAttribute(self, params):
+    def setMeshPolyhedraVertexAttribute(self, params):
         validate_schema(
             params,
-            self.mesh_polyhedrons_schemas_dict["vertex_attribute"],
-            self.mesh_polyhedrons_prefix,
+            self.mesh_polyhedra_schemas_dict["vertex_attribute"],
+            self.mesh_polyhedra_prefix,
         )
         id, name = params["id"], params["name"]
         self.displayAttributeOnVertices(id, name)
 
     @exportRpc(
-        mesh_polyhedrons_prefix
-        + mesh_polyhedrons_schemas_dict["polyhedron_attribute"]["rpc"]
+        mesh_polyhedra_prefix
+        + mesh_polyhedra_schemas_dict["polyhedron_attribute"]["rpc"]
     )
-    def setMeshPolyhedronsPolyhedronAttribute(self, params):
+    def setMeshPolyhedraPolyhedronAttribute(self, params):
         validate_schema(
             params,
-            self.mesh_polyhedrons_schemas_dict["polyhedron_attribute"],
-            self.mesh_polyhedrons_prefix,
+            self.mesh_polyhedra_schemas_dict["polyhedron_attribute"],
+            self.mesh_polyhedra_prefix,
         )
         id, name = params["id"], params["name"]
         self.displayAttributeOnCells(id, name)