+++ src/opengeodeweb_viewer/rpc/model/blocks/blocks_protocols.py
@@ -9,9 +9,9 @@ from opengeodeweb_viewer.utils_functions import get_schemas_dict, validate_schem
 from opengeodeweb_viewer.rpc.model.model_protocols import VtkModelView
 
 
-class VtkModelCornersPointsView(VtkModelView):
-    model_corners_points_prefix = "opengeodeweb_viewer.model.corners.points."
-    model_corners_points_schemas_dict = get_schemas_dict(
+class VtkModelBlocksView(VtkModelView):
+    model_blocks_prefix = "opengeodeweb_viewer.model.blocks."
+    model_blocks_schemas_dict = get_schemas_dict(
         os.path.join(os.path.dirname(__file__), "schemas")
     )
 
@@ -19,14 +19,14 @@ class VtkModelCornersPointsView(VtkModelView):
         super().__init__()
 
     @exportRpc(
-        model_corners_points_prefix
-        + model_corners_points_schemas_dict["visibility"]["rpc"]
+        model_blocks_prefix
+        + model_blocks_schemas_dict["visibility"]["rpc"]
     )
-    def setModelCornersPointsVisibility(self, params):
+    def setModelBlocksPolyhedraVisibility(self, params):
         validate_schema(
             params,
-            self.model_corners_points_schemas_dict["visibility"],
-            self.model_corners_points_prefix,
+            self.model_blocks_schemas_dict["visibility"],
+            self.model_blocks_prefix,
         )
         id, block_ids, visibility = (
             params["id"],
@@ -36,13 +36,14 @@ class VtkModelCornersPointsView(VtkModelView):
         self.SetBlocksVisibility(id, block_ids, visibility)
 
     @exportRpc(
-        model_corners_points_prefix + model_corners_points_schemas_dict["color"]["rpc"]
+        model_blocks_prefix
+        + model_blocks_schemas_dict["color"]["rpc"]
     )
-    def setModelCornersPointsColor(self, params):
+    def setModelBlocksPolyhedraColor(self, params):
         validate_schema(
             params,
-            self.model_corners_points_schemas_dict["color"],
-            self.model_corners_points_prefix,
+            self.model_blocks_schemas_dict["color"],
+            self.model_blocks_prefix,
         )
         id, block_ids, red, green, blue = (
             params["id"],