+++ src/opengeodeweb_viewer/rpc/model/surfaces/surfaces_protocols.py
@@ -9,9 +9,9 @@ from opengeodeweb_viewer.utils_functions import get_schemas_dict, validate_schem
 from opengeodeweb_viewer.rpc.model.model_protocols import VtkModelView
 
 
-class VtkModelSurfacesPolygonsView(VtkModelView):
-    model_surfaces_polygons_prefix = "opengeodeweb_viewer.model.surfaces.polygons."
-    model_surfaces_polygons_schemas_dict = get_schemas_dict(
+class VtkModelSurfacesView(VtkModelView):
+    model_surfaces_prefix = "opengeodeweb_viewer.model.surfaces."
+    model_surfaces_schemas_dict = get_schemas_dict(
         os.path.join(os.path.dirname(__file__), "schemas")
     )
 
@@ -19,14 +19,14 @@ class VtkModelSurfacesPolygonsView(VtkModelView):
         super().__init__()
 
     @exportRpc(
-        model_surfaces_polygons_prefix
-        + model_surfaces_polygons_schemas_dict["visibility"]["rpc"]
+        model_surfaces_prefix
+        + model_surfaces_schemas_dict["visibility"]["rpc"]
     )
     def setModelSurfacesPolygonsVisibility(self, params):
         validate_schema(
             params,
-            self.model_surfaces_polygons_schemas_dict["visibility"],
-            self.model_surfaces_polygons_prefix,
+            self.model_surfaces_schemas_dict["visibility"],
+            self.model_surfaces_prefix,
         )
         id, block_ids, visibility = (
             params["id"],
@@ -36,14 +36,14 @@ class VtkModelSurfacesPolygonsView(VtkModelView):
         self.SetBlocksVisibility(id, block_ids, visibility)
 
     @exportRpc(
-        model_surfaces_polygons_prefix
-        + model_surfaces_polygons_schemas_dict["color"]["rpc"]
+        model_surfaces_prefix
+        + model_surfaces_schemas_dict["color"]["rpc"]
     )
     def setModelSurfacesPolygonsCOlor(self, params):
         validate_schema(
             params,
-            self.model_surfaces_polygons_schemas_dict["color"],
-            self.model_surfaces_polygons_prefix,
+            self.model_surfaces_schemas_dict["color"],
+            self.model_surfaces_prefix,
         )
         id, block_ids, red, green, blue = (
             params["id"],