+++ src/opengeodeweb_viewer/vtk_protocol.py
@@ -30,12 +30,11 @@ class VtkView(vtk_protocols.vtkWebProtocol):
                 return p
 
     def render(self, view=-1):
-        renderer = self.get_renderer()
-        renderer_bounds = renderer.ComputeVisiblePropBounds()
-        grid_scale = self.get_object("grid_scale")["actor"]
-        grid_scale.SetBounds(renderer_bounds)
-        grid_scale.Modified()
-
+        if "grid_scale" in self.get_data_base():
+            renderer = self.get_renderer()
+            renderer_bounds = renderer.ComputeVisiblePropBounds()
+            grid_scale = self.get_object("grid_scale")["actor"]
+            grid_scale.SetBounds(renderer_bounds)
         self.get_protocol("vtkWebPublishImageDelivery").imagePush({"view": view})
 
     def register_object(self, id, reader, filter, actor, mapper, textures):